youtube/ai-worker/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.9.2
Pillow==11.0.0
